+++ hc/api/tests/test_notify_opsgenie.py
@@ -24,7 +24,7 @@ class NotifyOpsGenieTestCase(BaseTestCase):
         self.channel.save()
         self.channel.checks.add(self.check)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_opsgenie_with_legacy_value(self, mock_post):
         self._setup_data("123")
         mock_post.return_value.status_code = 202
@@ -39,7 +39,7 @@ class NotifyOpsGenieTestCase(BaseTestCase):
         payload = kwargs["json"]
         self.assertIn("DOWN", payload["message"])
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_opsgenie_up(self, mock_post):
         self._setup_data("123", status="up")
         mock_post.return_value.status_code = 202
@@ -53,7 +53,7 @@ class NotifyOpsGenieTestCase(BaseTestCase):
         method, url = args
         self.assertTrue(str(self.check.code) in url)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_opsgenie_with_json_value(self, mock_post):
         self._setup_data(json.dumps({"key": "456", "region": "eu"}))
         mock_post.return_value.status_code = 202
@@ -66,7 +66,7 @@ class NotifyOpsGenieTestCase(BaseTestCase):
         args, kwargs = mock_post.call_args
         self.assertIn("api.eu.opsgenie.com", args[1])
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_opsgenie_returns_error(self, mock_post):
         self._setup_data("123")
         mock_post.return_value.status_code = 403
@@ -76,7 +76,7 @@ class NotifyOpsGenieTestCase(BaseTestCase):
         n = Notification.objects.get()
         self.assertEqual(n.error, 'Received status code 403 with a message: "Nice try"')
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_handles_non_json_error_response(self, mock_post):
         self._setup_data("123")
         mock_post.return_value.status_code = 403