+++ hc/api/tests/test_notify_pushbullet.py
@@ -24,7 +24,7 @@ class NotifyPushbulletTestCase(BaseTestCase):
         self.channel.save()
         self.channel.checks.add(self.check)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_works(self, mock_post):
         mock_post.return_value.status_code = 200
 
@@ -38,7 +38,7 @@ class NotifyPushbulletTestCase(BaseTestCase):
         )
         self.assertEqual(kwargs["headers"]["Access-Token"], "fake-token")
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_escapes_body(self, mock_post):
         mock_post.return_value.status_code = 200
         self.check.name = "Foo & Bar"