+++ hc/api/tests/test_notify_pushover.py
@@ -26,7 +26,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         self.channel.save()
         self.channel.checks.add(self.check)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_works(self, mock_post):
         self._setup_data("123|0")
         mock_post.return_value.status_code = 200
@@ -45,7 +45,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         self.assertNotIn("All the other checks are up.", payload["message"])
         self.assertEqual(payload["tags"], self.check.unique_key)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_supports_up_priority(self, mock_post):
         self._setup_data("123|0|2", status="up")
         mock_post.return_value.status_code = 200
@@ -61,7 +61,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         self.assertIn("expire", payload)
 
     @override_settings(SECRET_KEY="test-secret")
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_obeys_rate_limit(self, mock_post):
         self._setup_data("123|0")
 
@@ -74,7 +74,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         n = Notification.objects.get()
         self.assertEqual(n.error, "Rate limit exceeded")
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_cancels_emergency_notification(self, mock_post):
         self._setup_data("123|2|0", status="up")
         mock_post.return_value.status_code = 200
@@ -92,7 +92,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         payload = up_kwargs["data"]
         self.assertIn("UP", payload["title"])
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_shows_all_other_checks_up_note(self, mock_post):
         self._setup_data("123|0")
         mock_post.return_value.status_code = 200
@@ -109,7 +109,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         payload = kwargs["data"]
         self.assertIn("All the other checks are up.", payload["message"])
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_lists_other_down_checks(self, mock_post):
         self._setup_data("123|0")
         mock_post.return_value.status_code = 200
@@ -128,7 +128,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         self.assertIn("Foobar", payload["message"])
         self.assertIn(other.cloaked_url(), payload["message"])
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_does_not_show_more_than_10_other_checks(self, mock_post):
         self._setup_data("123|0")
         mock_post.return_value.status_code = 200
@@ -147,7 +147,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         self.assertNotIn("Foobar", payload["message"])
         self.assertIn("11 other checks are also down.", payload["message"])
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_does_not_escape_title(self, mock_post):
         self._setup_data("123|0")
         self.check.name = "Foo & Bar"
@@ -160,7 +160,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         payload = kwargs["data"]
         self.assertEqual(payload["title"], "Foo & Bar is DOWN")
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_handles_disabled_priority(self, mock_post):
         self._setup_data("123|-3")
 
@@ -168,7 +168,7 @@ class NotifyPushoverTestCase(BaseTestCase):
         self.assertEqual(Notification.objects.count(), 0)
         self.assertFalse(mock_post.called)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_handles_disabled_up_priority(self, mock_post):
         self._setup_data("123|0|-3", status="up")