+++ hc/api/tests/test_notify_zulip.py
@@ -33,7 +33,7 @@ class NotifyZulipTestCase(BaseTestCase):
         self.channel.save()
         self.channel.checks.add(self.check)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_works(self, mock_post):
         mock_post.return_value.status_code = 200
 
@@ -51,7 +51,7 @@ class NotifyZulipTestCase(BaseTestCase):
         serialized = json.dumps(payload)
         self.assertNotIn(str(self.check.code), serialized)
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_returns_error(self, mock_post):
         mock_post.return_value.status_code = 403
         mock_post.return_value.json.return_value = {"msg": "Nice try"}
@@ -61,7 +61,7 @@ class NotifyZulipTestCase(BaseTestCase):
         n = Notification.objects.get()
         self.assertEqual(n.error, 'Received status code 403 with a message: "Nice try"')
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_handles_non_json_error_response(self, mock_post):
         mock_post.return_value.status_code = 403
         mock_post.return_value.json = Mock(side_effect=ValueError)
@@ -70,7 +70,7 @@ class NotifyZulipTestCase(BaseTestCase):
         n = Notification.objects.get()
         self.assertEqual(n.error, "Received status code 403")
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_uses_site_parameter(self, mock_post):
         mock_post.return_value.status_code = 200
         definition = {
@@ -100,7 +100,7 @@ class NotifyZulipTestCase(BaseTestCase):
         n = Notification.objects.get()
         self.assertEqual(n.error, "Zulip notifications are not enabled.")
 
-    @patch("hc.api.transports.requests.request")
+    @patch("hc.api.transports.requests.Session.request")
     def test_it_does_not_escape_topic(self, mock_post):
         mock_post.return_value.status_code = 200