+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 ### Improvements
 - Add support for EMAIL_USE_SSL environment variable (#685)
 
+### Bug Fixes
+- Fix the handling of TooManyRedirects exceptions
+
 ## v2.3 - 2022-08-05
 
 ### Improvements